Fixed tests that did not work anymore
