+++ tests/e2e-cypress/integration/infobox.cy.js
@@ -55,13 +55,16 @@ describe('The infobox', () => {
         })
 
         cy.get('[data-cy="infobox-content"]').then(($element) => {
-            const maxHeight = $element
+            const { paddingTop, paddingBottom } = getComputedStyle($element[0])
+            const verticalPadding = parseInt(paddingTop) + parseInt(paddingBottom)
+            const viewportHeight = Cypress.config('viewportHeight')
+            let maxHeight = $element
                 .find('[data-infobox="height-reference"]')
                 .toArray()
                 .map((child) => child.offsetHeight)
                 .reduce((max, height) => Math.max(max, height), 0)
-
-            expect($element.height()).to.be.equal(maxHeight)
+            maxHeight = Math.min(maxHeight + verticalPadding, viewportHeight * 0.35)
+            expect($element.height()).to.be.closeTo(maxHeight - verticalPadding, 0.1)
         })
     })
 })
